LibGfx/JPEGXL: Move modular transforms related code in its own file
No behavior change.
