+++ Userland/Libraries/LibGfx/CMakeLists.txt
@@ -64,6 +64,7 @@ set(SOURCES
     ImageFormats/JPEGLoader.cpp
     ImageFormats/JPEGXL/EntropyDecoder.cpp
     ImageFormats/JPEGXL/ICC.cpp
+    ImageFormats/JPEGXL/ModularTransforms.cpp
     ImageFormats/JPEGXL/SelfCorrectingPredictor.cpp
     ImageFormats/JPEGXLLoader.cpp
     ImageFormats/JPEGWriter.cpp
